Only install Sleeper if not already installed

+++ java/cdk-environment/src/main/resources/cloud-init.sh
@@ -53,9 +53,11 @@ sudo apt install -y docker-ce docker-ce-cli containerd.io binfmt-support qemu-us
 usermod -aG docker "$LOGIN_USER"
 
 # Install Sleeper CLI
-curl "https://raw.githubusercontent.com/gchq/sleeper/develop/scripts/cli/install.sh" -o "$LOGIN_HOME/sleeper-install.sh"
-chmod +x "$LOGIN_HOME/sleeper-install.sh"
-runuser --login "$LOGIN_USER" -c "$LOGIN_HOME/sleeper-install.sh"
+if [ ! -f "$LOGIN_HOME/.local/bin/sleeper"]; then
+  curl "https://raw.githubusercontent.com/gchq/sleeper/develop/scripts/cli/install.sh" -o "$LOGIN_HOME/sleeper-install.sh"
+  chmod +x "$LOGIN_HOME/sleeper-install.sh"
+  runuser --login "$LOGIN_USER" -c "$LOGIN_HOME/sleeper-install.sh"
+fi
 
 # Check out code
 REPOSITORY_DIR="$LOGIN_HOME/.sleeper/builder/$REPOSITORY"
